Refactor url and clean up url stuff
