+++ connect/src/com/telenor/connect/ui/ConnectLoginButton.java
@@ -7,29 +7,22 @@ import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
 import android.support.customtabs.CustomTabsCallback;
 import android.support.customtabs.CustomTabsClient;
 import android.support.customtabs.CustomTabsIntent;
 import android.support.customtabs.CustomTabsServiceConnection;
 import android.support.customtabs.CustomTabsSession;
-import android.text.TextUtils;
 import android.util.AttributeSet;
 import android.view.View;
 
 import com.telenor.connect.BrowserType;
-import com.telenor.connect.ConnectException;
 import com.telenor.connect.ConnectSdk;
 import com.telenor.connect.R;
-import com.telenor.connect.id.ConnectStore;
-import com.telenor.connect.utils.ClaimsParameterFormatter;
+import com.telenor.connect.utils.ConnectUrlHelper;
 import com.telenor.connect.utils.CustomTabsHelper;
 import com.telenor.connect.utils.Validator;
 
-import org.json.JSONException;
-
 import java.lang.ref.WeakReference;
-import java.util.HashMap;
 import java.util.Map;
 
 import static android.content.Intent.FLAG_ACTIVITY_NEW_TASK;
@@ -38,7 +31,7 @@ public class ConnectLoginButton extends ConnectWebViewLoginButton {
 
     private static final Uri PRE_FETCH_URL
             = Uri.parse(
-            ConnectSdk
+            ConnectUrlHelper
                     .getConnectApiUrl()
                     .newBuilder()
                     .addPathSegment("id")
@@ -53,12 +46,10 @@ public class ConnectLoginButton extends ConnectWebViewLoginButton {
     private boolean customTabsSupported = false;
     private boolean launchCustomTabInNewTask = true;
     private BrowserType browserType;
-    private ConnectStore connectStore;
     private CustomTabsSession session;
 
     public ConnectLoginButton(Context context, AttributeSet attributeSet) {
         super(context, attributeSet);
-        connectStore = new ConnectStore(getContext());
         setText(R.string.com_telenor_connect_login_button_text);
         connection = new WeakReferenceCustomTabsServiceConnection(new WeakReference<>(this));
         onClickListener = new LoginClickListener();
@@ -75,40 +66,7 @@ public class ConnectLoginButton extends ConnectWebViewLoginButton {
 
     private Uri getAuthorizeUri() {
         final Map<String, String> parameters = getParameters();
-        return ConnectSdk.getAuthorizeUri(parameters, browserType);
-    }
-
-    @NonNull
-    private Map<String, String> getParameters() {
-        final Map<String, String> parameters = new HashMap<>();
-        if (getAcrValues() != null && !getAcrValues().isEmpty()) {
-            parameters.put("acr_values", TextUtils.join(" ", getAcrValues()));
-        }
-
-        if (getLoginScopeTokens() != null && !getLoginScopeTokens().isEmpty()) {
-            parameters.put("scope", TextUtils.join(" ", getLoginScopeTokens()));
-        }
-
-        if (getClaims() != null && getClaims().getClaimsAsSet() != null) {
-            try {
-                parameters.put("claims", ClaimsParameterFormatter.asJson(getClaims()));
-            } catch (JSONException e) {
-                throw new ConnectException(
-                        "Failed to create claims Json. claims=" + getClaims(), e);
-            }
-        }
-
-        parameters.put("state", connectStore.generateSessionStateParam());
-        parameters.put("log_session_id", ConnectSdk.getLogSessionId());
-
-        if (!ConnectSdk.isCellularDataNetworkConnected()) {
-            parameters.put("prompt", "no_seam");
-        }
-
-        if (getLoginParameters() != null && !getLoginParameters().isEmpty()) {
-            parameters.putAll(getLoginParameters());
-        }
-        return parameters;
+        return ConnectUrlHelper.getAuthorizeUri(parameters, browserType);
     }
 
     private static boolean contextIntentFilterMatchesRedirectUri(Context context) {
@@ -142,13 +100,7 @@ public class ConnectLoginButton extends ConnectWebViewLoginButton {
     }
 
     private void launchWebViewAuthentication() {
-        int customLoadingLayout = getCustomLoadingLayout();
-        if (customLoadingLayout == NO_CUSTOM_LAYOUT) {
-            ConnectSdk.authenticate(getActivity(), getParameters(), getRequestCode());
-        } else {
-            ConnectSdk.authenticate(
-                    getActivity(), getParameters(), customLoadingLayout, getRequestCode());
-        }
+        startWebViewAuthentication();
     }
 
     private void launchChromeCustomTabAuthentication() {
@@ -175,7 +127,6 @@ public class ConnectLoginButton extends ConnectWebViewLoginButton {
         @Override
         public void onClick(View v) {
             Validator.sdkInitialized();
-
             ConnectSdk.beforeAuthentication();
 
             if (!customTabsSupported) {
