+++ connect/src/com/telenor/connect/ui/ConnectWebViewLoginButton.java
@@ -1,6 +1,7 @@
 package com.telenor.connect.ui;
 
 import android.content.Context;
+import android.support.annotation.NonNull;
 import android.text.TextUtils;
 import android.util.AttributeSet;
 import android.view.View;
@@ -9,8 +10,8 @@ import com.telenor.connect.ConnectException;
 import com.telenor.connect.ConnectSdk;
 import com.telenor.connect.R;
 import com.telenor.connect.id.Claims;
+import com.telenor.connect.id.ConnectStore;
 import com.telenor.connect.utils.ClaimsParameterFormatter;
-import com.telenor.connect.utils.ConnectUtils;
 import com.telenor.connect.utils.Validator;
 
 import org.json.JSONException;
@@ -22,7 +23,7 @@ import java.util.Map;
 
 public class ConnectWebViewLoginButton extends ConnectButton {
 
-    protected static final int NO_CUSTOM_LAYOUT = -1;
+    public static final int NO_CUSTOM_LAYOUT = -1;
 
     private ArrayList<String> acrValues;
     private Map<String, String> loginParameters;
@@ -30,11 +31,13 @@ public class ConnectWebViewLoginButton extends ConnectButton {
     private int requestCode = 0xa987;
     private Claims claims;
     private int customLoadingLayout = NO_CUSTOM_LAYOUT;
+    private ConnectStore connectStore;
 
     public ConnectWebViewLoginButton(Context context, AttributeSet attributeSet) {
         super(context, attributeSet);
         setText(R.string.com_telenor_connect_login_button_text);
         setOnClickListener(new LoginClickListener());
+        connectStore = new ConnectStore(getContext());
     }
 
     public ArrayList<String> getAcrValues() {
@@ -93,44 +96,63 @@ public class ConnectWebViewLoginButton extends ConnectButton {
         this.customLoadingLayout = customLoadingLayout;
     }
 
+    protected void startWebViewAuthentication() {
+        ConnectSdk.authenticate(getActivity(), getParameters(), getCustomLoadingLayout(), getRequestCode());
+    }
+
+    @NonNull
+    protected Map<String, String> getParameters() {
+        Map<String, String> parameters = new HashMap<>();
+        if (getAcrValues() != null && !getAcrValues().isEmpty()) {
+            parameters.put("acr_values", TextUtils.join(" ", getAcrValues()));
+        }
+
+        if (getLoginScopeTokens() != null && !getLoginScopeTokens().isEmpty()) {
+            parameters.put("scope", TextUtils.join(" ", getLoginScopeTokens()));
+        }
+
+        addClaims(parameters);
+
+        if (getLoginParameters() != null && !getLoginParameters().isEmpty()) {
+            parameters.putAll(getLoginParameters());
+        }
+
+        handleStatePromptAndLogSessionId(parameters);
+        return parameters;
+    }
+
+    private void addClaims(Map<String, String> parameters) {
+        if (getClaims() != null && getClaims().getClaimsAsSet() != null) {
+            try {
+                parameters.put("claims", ClaimsParameterFormatter.asJson(getClaims()));
+            } catch (JSONException e) {
+                throw new ConnectException(
+                        "Failed to create claims Json. claims=" + getClaims(), e);
+            }
+        }
+    }
+
+    private void handleStatePromptAndLogSessionId(Map<String, String> parameters) {
+        if (TextUtils.isEmpty(parameters.get("state"))) {
+            parameters.put("state", connectStore.generateSessionStateParam());
+        }
+        if (TextUtils.isEmpty(parameters.get("prompt")) && !ConnectSdk.isCellularDataNetworkConnected()) {
+            parameters.put("prompt", "no_seam");
+        }
+        if (TextUtils.isEmpty(parameters.get("log_session_id"))) {
+            parameters.put("log_session_id", ConnectSdk.getLogSessionId());
+        }
+    }
+
+
     private class LoginClickListener implements OnClickListener {
 
         @Override
         public void onClick(View v) {
             Validator.sdkInitialized();
-
             ConnectSdk.beforeAuthentication();
-
-            Map<String, String> parameters = new HashMap<>();
-            if (getAcrValues() != null && !getAcrValues().isEmpty()) {
-                parameters.put(ConnectUtils.ACR_VALUES_PARAM_NAME,
-                        TextUtils.join(" ", getAcrValues()));
-            }
-
-            if (getLoginScopeTokens() != null && !getLoginScopeTokens().isEmpty()) {
-                parameters.put("scope", TextUtils.join(" ", getLoginScopeTokens()));
-            }
-
-            if (claims != null && claims.getClaimsAsSet() != null) {
-                try {
-                    parameters.put("claims", ClaimsParameterFormatter.asJson(claims));
-                } catch (JSONException e) {
-                    throw new ConnectException("Failed to create claims Json. claims="+claims, e);
-                }
-            }
-
-            if (getLoginParameters() != null && !getLoginParameters().isEmpty()){
-                parameters.putAll(getLoginParameters());
-            }
-
-            if (customLoadingLayout == NO_CUSTOM_LAYOUT) {
-                ConnectSdk.authenticate(getActivity(), parameters, requestCode);
-            } else {
-                ConnectSdk.authenticate(getActivity(),
-                        parameters,
-                        customLoadingLayout,
-                        requestCode);
-            }
+            startWebViewAuthentication();
         }
+
     }
 }
