+++ connect/tests/src/test/java/com/telenor/connect/ConnectSdkTest.java
@@ -4,6 +4,7 @@ import android.content.Intent;
 import android.net.Uri;
 
 import com.telenor.connect.id.ConnectIdService;
+import com.telenor.connect.utils.ConnectUrlHelper;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -30,7 +31,7 @@ public class ConnectSdkTest {
 
     @Test
     public void getConnectApiUrlReturnsProductionByDefault() {
-        assertThat(ConnectSdk.getConnectApiUrl().toString(),
+        assertThat(ConnectUrlHelper.getConnectApiUrl().toString(),
                 is("https://connect.telenordigital.com/"));
     }
 
@@ -50,7 +51,7 @@ public class ConnectSdkTest {
         HashMap<String, String> parameters = new HashMap<>();
         parameters.put("state", "xyz");
         parameters.put("scope", "anything");
-        ConnectSdk.getAuthorizeUri(parameters, null);
+        ConnectUrlHelper.getAuthorizeUri(parameters, null);
 
         Intent intent = new Intent();
         intent.setData(Uri.parse("connect-tests://oauth2callback?state=xyz&code=abc"));
@@ -64,7 +65,7 @@ public class ConnectSdkTest {
         HashMap<String, String> parameters = new HashMap<>();
         parameters.put("state", "xyz");
         parameters.put("scope", "anything");
-        ConnectSdk.getAuthorizeUri(parameters, null);
+        ConnectUrlHelper.getAuthorizeUri(parameters, null);
 
         Intent intent = new Intent();
         intent.setData(Uri.parse("something-not-registed://oauth2callback?state=xyz&code=abc"));
@@ -78,7 +79,7 @@ public class ConnectSdkTest {
         HashMap<String, String> parameters = new HashMap<>();
         parameters.put("state", "xyz");
         parameters.put("scope", "anything");
-        ConnectSdk.getAuthorizeUri(parameters, null);
+        ConnectUrlHelper.getAuthorizeUri(parameters, null);
 
         Intent intent = new Intent();
         intent.setData(Uri.parse("something-not-registed://oauth2callback?state=NNN&code=abc"));
@@ -92,7 +93,7 @@ public class ConnectSdkTest {
         HashMap<String, String> parameters = new HashMap<>();
         parameters.put("state", "xyz");
         parameters.put("scope", "anything");
-        ConnectSdk.getAuthorizeUri(parameters, null);
+        ConnectUrlHelper.getAuthorizeUri(parameters, null);
 
         Intent intent = new Intent();
         intent.setData(Uri.parse("something-not-registed://oauth2callback?state=xyz"));
